Moved "Next Aurax" click event logic from weapons.js to weapons-table.js
