+++ priv/static/js/character/weapons-table.js
@@ -1,3 +1,5 @@
+import { nextAuraxElementID } from "/js/character/weapons.js";
+
 window.addEventListener('load', (event) => {
     if (window.innerWidth >= 768) {
         $("input.search-input:first").focus();
@@ -6,6 +8,27 @@ window.addEventListener('load', (event) => {
     //TODO edit dataset based on cached persistent data
     //document.getElementById("weaponTable").dataset.pagination = false;
 });
+
+function initializeButtonEvent() {
+    document.getElementById("nextAurax").addEventListener('click', function () {
+        $('html, body').animate({
+            scrollTop: $("#" + nextAuraxElementID).offset().top - 300 //- 254 to be at top
+        }, 500);
+        setTimeout(function () {
+            flashElement(nextAuraxElementID);
+        }, 500);
+    });
+}
+
+function flashElement(elementId) {
+    let flashInterval = setInterval(function () {
+        $("#" + elementId).toggleClass("flash-border");
+    }, 250);
+    setTimeout(function () {
+        window.clearInterval(flashInterval);
+        $("#" + elementId).removeClass("flash-border");
+    }, 1000);
+}
 
 export default function init() {
     $('#weaponTable').bootstrapTable({
@@ -22,4 +45,6 @@ export default function init() {
             return template.content.querySelector(".weaponName").innerHTML.toLowerCase().indexOf(text.toLowerCase()) > -1
         })
     }
+
+    initializeButtonEvent();
 }
